+++ src/qq_lib/submit/factory.py
@@ -4,8 +4,6 @@
 from dataclasses import fields
 from pathlib import Path
 
-from click import Parameter
-
 from qq_lib.batch.interface import BatchInterface, BatchMeta
 from qq_lib.core.common import split_files_list
 from qq_lib.core.error import QQError
@@ -24,19 +22,18 @@ class SubmitterFactory:
     the command-line and from the script itself.
     """
 
-    def __init__(
-        self, script: Path, params: list[Parameter], command_line: list[str], **kwargs
-    ):
+    def __init__(self, script: Path, command_line: list[str], **kwargs):
         """
         Initialize the factory with the script, command-line parameters, and additional options.
 
         Args:
             script (Path): Path to the script to submit.
-            params (list[Parameter]): List of all known submission parameters.
             command_line (list[str]): All the arguments and options specified on the command line.
             **kwargs: Keyword arguments from the command line.
         """
-        self._parser = Parser(script, params)
+        from qq_lib.submit.cli import submit
+
+        self._parser = Parser(script, submit.params)
         self._script = script
         self._input_dir = script.parent
         self._kwargs = kwargs